renamed Map component
